GDP-14 cambiar vista fechas director
